% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Accident.R
\name{build_graph_Event_Accident}
\alias{build_graph_Event_Accident}
\title{Event - Accident}
\usage{
build_graph_Event_Accident(Rgraph)
}
\arguments{
\item{Rgraph}{The Rgraph object (package specific object to save the graph)}
}
\value{
Rgraph
}
\description{
build the nodes and edges in the graph object necessary to include the Loss Aggregator
}
