% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Death.R
\name{build_graph_Event_Death}
\alias{build_graph_Event_Death}
\title{Event - Death}
\usage{
build_graph_Event_Death(Rgraph)
}
\arguments{
\item{Rgraph}{The Rgraph object (package specific object to save the graph)}
}
\value{
Rgraph
}
\description{
build the nodes and edges in the graph object necessary to include the Loss Aggregator
}
