% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impact_CoC.R
\name{build_graph_Impact_CoC}
\alias{build_graph_Impact_CoC}
\title{Impact -  Cost of Care}
\usage{
build_graph_Impact_CoC(Rgraph)
}
\arguments{
\item{Rgraph}{The Rgraph object (package specific object to save the graph)}
}
\value{
Rgraph
}
\description{
build the nodes and edges in the graph object necessary to include the Loss Aggregator
}
