% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impact_StockDevaluation.R
\name{build_graph_Impact_StockDevaluation}
\alias{build_graph_Impact_StockDevaluation}
\title{Impact - Stock Devaluation}
\usage{
build_graph_Impact_StockDevaluation(Rgraph)
}
\arguments{
\item{Rgraph}{The Rgraph object (package specific object to save the graph)}
}
\value{
Rgraph
}
\description{
build the nodes and edges in the graph object necessary to include the Loss Aggregator
}
