% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monetary_Loss.R
\name{build_graph_Monetary_Loss}
\alias{build_graph_Monetary_Loss}
\title{Loss Aggregator}
\usage{
build_graph_Monetary_Loss(Rgraph)
}
\arguments{
\item{Rgraph}{The Rgraph object (package specific object to save the graph)}
}
\value{
Rgraph
}
\description{
build the nodes and edges in the graph object necessary to include the Loss Aggregator
}
