% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Accident.R
\name{Event_Accident}
\alias{Event_Accident}
\title{Event - Accident}
\usage{
Event_Accident(input_name, input)
}
\arguments{
\item{input_name}{currently only 'Data_none' and 'Data_bool_War' are supported.}

\item{input}{e.g. bool_war}
}
\value{
per_disability
}
\description{
calculates the probability of an accident.
}
