% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Illness.R
\name{Event_Ilness}
\alias{Event_Ilness}
\title{Event - Illness}
\usage{
Event_Ilness(input_name, input)
}
\arguments{
\item{input_name}{nothing}

\item{input}{nothing}
}
\value{
per_disability
}
\description{
calculates the probability of an Illness. Not implemented yet. Just a placeholder
}
