% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_StockCrash.R
\name{Event_StockCrash}
\alias{Event_StockCrash}
\title{Event - Stock Market Crash}
\usage{
Event_StockCrash(input_name, input)
}
\arguments{
\item{input_name}{nothing}

\item{input}{nothing}
}
\value{
Data_per_StockLoss
}
\description{
calculates the probability of a stock crash. Not implemented yet. Just a placeholder
}
