% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Unemployment.R
\name{Event_Unemployment}
\alias{Event_Unemployment}
\title{Event - Unemployment}
\usage{
Event_Unemployment(input_name, input)
}
\arguments{
\item{input_name}{nothing}

\item{input}{nothing}
}
\value{
Data_per_unemployment
}
\description{
calculates the probability of a unemplyoment. Not implemented yet. Just a placeholder
}
