% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_War.R
\name{Event_War}
\alias{Event_War}
\title{Event - War}
\usage{
Event_War()
}
\arguments{
\item{input_name}{currently only 'none' is supported}

\item{input}{currently unused}
}
\value{
bool_War
}
\description{
calculates the probability of a war.
}
