% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Build.R
\name{Graph_Build}
\alias{Graph_Build}
\title{Build causality Graph for RisksOfLife}
\usage{
Graph_Build()
}
\value{
Rgraph The object that represents the graph structure / the causal chain
}
\description{
This will hopefully be replaced by a load functionality on existing graphs, together with a more intuitive way to create the graphs.
}
