% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_calculatePaths.R
\name{Graph_calculate_paths}
\alias{Graph_calculate_paths}
\title{calculate given paths}
\usage{
Graph_calculate_paths()
}
\arguments{
\item{paths}{The paths object representing the causality chains}
}
\value{
path_results
}
\description{
calculate given paths
}
