% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Build.R
\name{Graph_convert_to_igraph}
\alias{Graph_convert_to_igraph}
\title{Convert the RiskOfLife specific graph format to igraph}
\usage{
Graph_convert_to_igraph(Rgraph, onlyActive)
}
\arguments{
\item{Rgraph}{the Rgraph RoL object}

\item{onlyActive}{only consider active relations}
}
\value{
igraph The object that represents the graph structure / the causal chain
}
\description{
Convert the RiskOfLife specific graph format to igraph
}
