% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Build.R
\name{Graph_convert_to_igraph_edges}
\alias{Graph_convert_to_igraph_edges}
\title{Convert the RiskOfLife specific graph format to an igraph compatbile edges dataframe}
\usage{
Graph_convert_to_igraph_edges(Rgraph, onlyActive)
}
\arguments{
\item{Rgraph}{the Rgraph RoL object}

\item{onlyActive}{only consider active relations}
}
\value{
edges data frame that should be compatible with igraph's graph_from_data_frame function
}
\description{
Convert the RiskOfLife specific graph format to an igraph compatbile edges dataframe
}
