% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_convertFormat.R
\name{Graph_convert_to_igraph_nodes}
\alias{Graph_convert_to_igraph_nodes}
\title{Convert the RiskOfLife specific graph format to a igraph compatible node dataframe}
\usage{
Graph_convert_to_igraph_nodes(Rgraph, onlyActive = TRUE)
}
\arguments{
\item{Rgraph}{the Rgraph RoL object}

\item{onlyActive}{only consider active relations}
}
\value{
Nodes  data frame that should be compatible with igraph's graph_from_data_frame function
}
\description{
Convert the RiskOfLife specific graph format to a igraph compatible node dataframe
}
