% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_getAllPaths.R
\name{Graph_getAllPaths}
\alias{Graph_getAllPaths}
\title{Get all paths from RiskOfLife graph object}
\usage{
Graph_getAllPaths(Rgraph, startNode, endNode, onlyActive)
}
\arguments{
\item{Rgraph}{The Rgraph object (package specific object to save the graph)}
}
\value{
Rgraph The object that represents the graph structure / the causal chain
}
\description{
Currently no circular paths are allowed, e.g. in one path the same node cannot be hit more than once.
This might be subject to change. Currently I assume such circular dependencies are not relevant in the overall calculation as their probability is extremly small.
}
