% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_getResultAtNode.R
\name{Graph_getResultAtNode}
\alias{Graph_getResultAtNode}
\title{calculate the overall distribution at a given node for a given set of paths}
\usage{
Graph_getResultAtNode(paths, path_results, node)
}
\arguments{
\item{paths}{The paths object representing the causality chains}

\item{node}{At which node should the result be calculated}
}
\value{
path_results
}
\description{
calculate the overall distribution at a given node for a given set of paths
}
