% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Validation.R
\name{Graph_Validation}
\alias{Graph_Validation}
\title{Validate the Neo4j data used for RisksOfLife}
\usage{
Graph_Validation(graph)
}
\arguments{
\item{graph}{The graph object representing the neo4j database as produced by Rneo4j}
}
\value{
nothing
}
\description{
This function will give errors or warning depent on urgency.
It checks the graph for being a plausible input for RisksOfLife.
}
