% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impact_CoC.R
\name{Impact_CoC}
\alias{Impact_CoC}
\title{Impact -  Cost of Care}
\usage{
Impact_CoC(input_name, input)
}
\arguments{
\item{input_name}{currently only 'Data_per_disability' is supported}

\item{input}{the input data, e.g. the variable 'per_disability'}
}
\value{
data_monetaryLoss
}
\description{
given the natural units input it calculates the  the monetary loss.
}
