% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impact_LoE.R
\name{Impact_LoE}
\alias{Impact_LoE}
\title{Impact - Loss of Earnings}
\usage{
Impact_LoE(input_name, input)
}
\arguments{
\item{input_name}{currently only 'Data_bool_dead' is supported}

\item{input}{the input data, e.g. the variable 'Data_bool_dead'}
}
\value{
data_monetaryLoss
}
\description{
given the natural units input it calculates the monetary loss.
}
