% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impact_StockDevaluation.R
\name{Impact_StockDevaluation}
\alias{Impact_StockDevaluation}
\title{Impact - Stock Devaluation}
\usage{
Impact_StockDevaluation(input_name, input)
}
\arguments{
\item{input_name}{nothing}

\item{input}{nothing}
}
\value{
data_monetaryLoss
}
\description{
given the natural units input it calculates the monetary loss. Not implemented yet. Just a placeholder
}
