% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monetary_Loss.R
\name{Monetary_Loss}
\alias{Monetary_Loss}
\title{Loss Aggregator}
\usage{
Monetary_Loss(data_monetaryLoss)
}
\arguments{
\item{data_monetaryLoss}{This is an object containing frequency and severity of monetary losses.}
}
\value{
monetary, aggregated Loss
}
\description{
not functional yet. the intention is to have this as the last node for all paths. This node should then aggregate the results.
}
