% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_convertFormat.R
\name{Paths_convert_to_paths_edges_map}
\alias{Paths_convert_to_paths_edges_map}
\title{Convert the RiskOfLife specific Paths format to a map from paths to edges}
\usage{
Paths_convert_to_paths_edges_map(paths, Edges)
}
\arguments{
\item{Paths}{the Paths RoL object}
}
\value{
table that has paths and corresponding edges
}
\description{
Convert the RiskOfLife specific Paths format to a map from paths to edges
}
