% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_convertFormat.R
\name{Paths_convert_to_paths_nodes_map}
\alias{Paths_convert_to_paths_nodes_map}
\title{Convert the RiskOfLife specific Paths format to a map from paths to nodes}
\usage{
Paths_convert_to_paths_nodes_map(paths)
}
\arguments{
\item{Paths}{the Paths RoL object}
}
\value{
table that has paths and corresponding nodes
}
\description{
Convert the RiskOfLife specific Paths format to a map from paths to nodes
}
