% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_RoL.R
\name{server}
\alias{server}
\title{Draw Rshiny Graph}
\usage{
server(input, output)
}
\arguments{
\item{input}{input}

\item{output}{output}
}
\value{
nothing
}
\description{
This is to visualize the Risk of Life Graph in Rshiny
}
