% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_bool_dead.R
\name{Val_Data_bool_dead}
\alias{Val_Data_bool_dead}
\title{Data Type - bool_dead}
\usage{
Val_Data_bool_dead(bool_dead)
}
\arguments{
\item{bool_dead}{put in the object you want to check.}
}
\value{
none the code will just throw an error if the object is not as expected.
}
\description{
verifies wether the object has the expected format and fullfills some boundary conditions.
}
