% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_bool_War.R
\name{Val_Data_bool_War}
\alias{Val_Data_bool_War}
\title{Data Type - bool_war}
\usage{
Val_Data_bool_War(bool_war)
}
\arguments{
\item{bool_war}{put in the object you want to check.}
}
\value{
none the code will just throw an error if the object is not as expected.
}
\description{
verifies wether the object has the expected format and fullfills some boundary conditions.
}
