% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_none.R
\name{Val_Data_none}
\alias{Val_Data_none}
\title{Data Type - per_disability}
\usage{
Val_Data_none(none)
}
\arguments{
\item{per_disability}{put in the object you want to check.}
}
\value{
none the code will just throw an error if the object is not as expected.
}
\description{
verifies wether the object has the expected format and fullfills some boundary conditions. This is just a dummy that makes my life easier, as with it I can more easily query for the codes for all data types in the Graph Validation.
}
