% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_per_disability.R
\name{Val_Data_per_disability}
\alias{Val_Data_per_disability}
\title{Data Type - per_disability}
\usage{
Val_Data_per_disability(per_disability)
}
\arguments{
\item{per_disability}{put in the object you want to check.}
}
\value{
none the code will just throw an error if the object is not as expected.
}
\description{
verifies wether the object has the expected format and fullfills some boundary conditions.
}
