% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_per_monetaryLoss.R
\name{Val_Data_per_monetaryLoss}
\alias{Val_Data_per_monetaryLoss}
\title{Data Type - per_monetaryLoss}
\usage{
Val_Data_per_monetaryLoss(per_monetaryLoss)
}
\arguments{
\item{per_monetaryLoss}{put in the object you want to check.}
}
\value{
none the code will just throw an error if the object is not as expected.
}
\description{
verifies wether the object has the expected format and fullfills some boundary conditions.
}
