% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_per_StockLoss.R
\name{Val_Data_per_StockLoss}
\alias{Val_Data_per_StockLoss}
\title{Data Type - per_StockLoss}
\usage{
Val_Data_per_StockLoss(per_StockLoss)
}
\arguments{
\item{per_StockLoss}{put in the object you want to check.}
}
\value{
none the code will just throw an error if the object is not as expected.
}
\description{
verifies wether the object has the expected format and fullfills some boundary conditions.
}
