% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_per_unemployment.R
\name{Val_Data_per_unemployment}
\alias{Val_Data_per_unemployment}
\title{Data Type - per_unemployment}
\usage{
Val_Data_per_unemployment(per_unemployment)
}
\arguments{
\item{per_unemployment}{put in the object you want to check.}
}
\value{
none the code will just throw an error if the object is not as expected.
}
\description{
verifies wether the object has the expected format and fullfills some boundary conditions.
}
