% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Build.R
\name{Graph_Build}
\alias{Graph_Build}
\title{Build the Neo4j data used for RisksOfLife}
\usage{
Graph_Build(graph)
}
\arguments{
\item{graph}{The graph object representing the neo4j database as produced by Rneo4j}
}
\value{
nothing
}
\description{
This will hopefully be replaced by a load functionality on existing graphs, together with a more intuitive way to create the graphs.
}
