% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_getAllPaths.R
\name{Graph_getAllPaths}
\alias{Graph_getAllPaths}
\title{Get all paths from RiskOfLife graph object}
\usage{
Graph_getAllPaths(Rgraph, startNode, endNode, onlyActive)
}
\arguments{
\item{Rgraph}{The Rgraph object (package specific object to save the graph)}
}
\value{
Rgraph The object that represents the graph structure / the causal chain
}
\description{
This will hopefully be replaced by a load functionality on existing graphs, together with a more intuitive way to create the graphs.
}
