% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{RisksOfLife_main}
\alias{RisksOfLife_main}
\title{Main R file to show the dataflow of RisksOfLife}
\usage{
RisksOfLife_main()
}
\value{
path_results
}
\description{
This function is the main function for RisksOfLife. not sure if I will keep it. I used it to check whether the whole concept actually calculates plausible results.
It will likely be replaced by jaavascript that then directly accesses the "Graph*"-functions
}
